refactor: tests
