+++ src/tests/load-purejs-crypto.ts
@@ -1,0 +1,4 @@
+import { setCryptoImplementation } from '../crypto/index.ts'
+import { pureJsCrypto } from '../crypto/pure.ts'
+
+setCryptoImplementation(pureJsCrypto)
