+++ src/tests/load-webcrypto.ts
@@ -1,0 +1,4 @@
+import { setCryptoImplementation } from '../crypto/index.ts'
+import { webcryptoCrypto } from '../crypto/webcrypto.ts'
+
+setCryptoImplementation(webcryptoCrypto)
