+++ src/tests/utils.ts
@@ -1,15 +1,12 @@
-import P from 'pino'
+import assert from 'node:assert'
+import { pino } from 'pino'
 
 export function delay(ms: number) {
 	return new Promise((resolve) => setTimeout(resolve, ms))
 }
 
 export function expectBuffsEq(a: Uint8Array, b: Uint8Array) {
-	expect(
-		Array.from(a)
-	).toEqual(
-		Array.from(b)
-	)
+	assert.deepEqual(Array.from(a), Array.from(b))
 }
 
 /**
@@ -20,5 +17,5 @@ export function bufferFromHexStringWithWhitespace(txt: string) {
 	return Buffer.from(txt.replace(/\s/g, ''), 'hex')
 }
 
-export const logger = P({})
+export const logger = pino({})
 logger.level = process.env.LOG_LEVEL || 'info'
